Reorganize @subsquid/squid-sdk subpaths; extract fallback core; optional EVM peers
- Hierarchical subpaths replacing flat old-name slugs: evm, evm/rpc,
  evm/fallback, solana, fuel, starknet, fallback, util, util/<x>, client/<x>,
  logger.
- Extract the VM-agnostic fallback supervisor into its own `fallback` subpath;
  the EVM binding (createEvmFallbackSource) lives at `evm/fallback`.
- Make @subsquid/evm-rpc and @subsquid/evm-normalization optional peer
  dependencies. The fallback core has no EVM in its runtime graph, and
  evm/fallback requires evm/rpc lazily, so a Portal-only/non-EVM squid needs
  neither peer.

exports + typesVersions updated for the new tree; verified type-checking under
classic (template) and nodenext resolution and runtime require() for all 19
subpaths.
