+++ sdk/squid-sdk/src/evm/fallback/index.ts
@@ -1,0 +1,1 @@
+export * from './source'
