+++ sdk/squid-sdk/src/evm/rpc/decode/decode.lfs.test.ts
