+++ sdk/squid-sdk/src/evm/rpc/networks/base-mainnet.ts
