+++ sdk/squid-sdk/src/evm/rpc/networks/ethereum-mainnet.ts
