+++ sdk/squid-sdk/src/evm/rpc/networks/gnosis-mainnet.ts
