+++ sdk/squid-sdk/src/evm/rpc/networks/networks.test.ts
