+++ sdk/squid-sdk/src/fallback/capability.test.ts
