+++ sdk/squid-sdk/src/fallback/capability.ts
