+++ sdk/squid-sdk/src/fallback/diagnostics.test.ts
