+++ sdk/squid-sdk/src/fallback/diagnostics.ts
