+++ sdk/squid-sdk/src/fallback/health.test.ts
