+++ sdk/squid-sdk/src/fallback/health.ts
