+++ sdk/squid-sdk/src/fallback/index.ts
@@ -4,5 +4,4 @@ export * from './health'
 export * from './selector'
 export * from './capability'
 export * from './fallback'
-export * from './evm/source'
 export * from './metrics'
