+++ sdk/squid-sdk/src/fallback/selector.ts
