+++ sdk/squid-sdk/src/fuel/index.ts
