+++ sdk/squid-sdk/src/util/hex/index.ts
