+++ sdk/squid-sdk/src/util/validation/index.ts
